+++ config/lottery-tools.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+    'bbfs' => [
+        'minimum_unique_digits' => 2,
+        'maximum_unique_digits' => 7,
+        'allowed_output_lengths' => [2, 3, 4],
+    ],
+];